Split client registry behavior tests
